backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
langgraph==0.2.14
langchain==0.2.12
langchain-google-genai==1.0.7
langchain-community==0.2.10
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
requests==2.32.3
beautifulsoup4==4.12.3
redis==5.0.7
supabase==2.5.3
pydantic==2.8.2
python-dotenv==1.0.1
sse-starlette==2.1.2
httpx==0.27.0
aiohttp==3.9.5
duckduckgo-search==8.1.1
